Split out components from requirements/requirements.js
