+++ ui/__tests__/components/requirements/policy-link.test.js
@@ -1,0 +1,13 @@
+import { shallow } from 'enzyme';
+import React from 'react';
+
+import PolicyLink from '../../../components/requirements/policy-link';
+
+describe('<PolicyLink />', () => {
+  it('renders using title_with_number from the policy prop', () => {
+    const policy = { id: 10, title_with_number: 'fake-text' };
+    const rendered = shallow(<PolicyLink policy={policy} />);
+    const link = rendered.find('a');
+    expect(link.text()).toEqual('fake-text');
+  });
+});
